Variable type has been changed.

+++ saga-choreography/inventory-service/src/main/java/com/market/inventory/application/core/usecase/FindInventoryByProductIdUseCase.java
@@ -13,7 +13,7 @@ public class FindInventoryByProductIdUseCase implements FindInventoryByProductId
     }
 
     @Override
-    public Inventory find(Integer productId) {
+    public Inventory find(Long productId) {
 
         return findInventoryByProductIdOutputPort.find(productId).orElseThrow(
                 () -> new RuntimeException("Not found inventory by this product.")
